Fix wheezy → jessie upgrade failure when running under systemd
During the upgrade the statically shipped symlink is replaced by one
which is created in postinst by init-system-helpers.
When systemd is reloaded while the symlink is missing, it loses track of
the running rsyslogd process. As a workaround, create a runtime copy of
the syslog.service symlink before the upgrade and remove it afterwards.
See #724796